+++ packages/waggle/src/constants.ts
@@ -1,8 +1,8 @@
-import path from "path";
+import path from 'path'
 import os from 'os'
 
 export enum Config {
-    ZBAY_DIR = '.zbay'
+  ZBAY_DIR = '.zbay'
 }
 
 export const ZBAY_DIR_PATH = path.join(os.homedir(), Config.ZBAY_DIR)