+++ packages/waggle/src/sleep.ts
@@ -1,5 +1,5 @@
-export const sleep = (time = 1000) =>
-  new Promise<void>(resolve => {
+export const sleep = async (time = 1000) =>
+  await new Promise<void>(resolve => {
     setTimeout(() => {
       resolve()
     }, time)