+++ packages/waggle/src/socket/constants.ts
@@ -8,5 +8,14 @@ export enum EventTypesServer {
   SUBSCRIBE_FOR_TOPIC = 'subscribeForTopic',
   ADD_TOR_SERVICE = 'addTorService',
   REMOVE_TOR_SERVICE = 'removeTorService',
-  GET_PUBLIC_CHANNELS = 'getPublicChannels'
+  GET_PUBLIC_CHANNELS = 'getPublicChannels',
+  INITIALIZE_CONVERSATION = 'initializeConversation',
+  GET_AVAILABLE_USERS = 'getAvailableUsers',
+  GET_PRIVATE_CONVERSATIONS = 'getPrivateConversations',
+  SEND_PRIVATE_MESSAGE = 'sendPrivateMessage',
+  ADD_USER = 'addUser',
+  SEND_DIRECT_MESSAGE = 'sendDirectMessage',
+  FETCH_ALL_DIRECT_MESSAGES = 'fetchAllDirectMessages',
+  SUBSCRIBE_FOR_DIRECT_MESSAGE_THREAD = 'subscribeForDirectMessageThread',
+  DIRECT_MESSAGE = 'directMessage'
 }