+++ packages/waggle/src/socket/constantsReponse.ts
@@ -3,5 +3,8 @@ export enum EventTypesResponse {
   RESPONSE_FETCH_ALL_MESSAGES = 'responseFetchAllMessages',
   RESPONSE_ADD_TOR_SERVICE = 'responseAddTorService',
   RESPONSE_REMOVE_TOR_SERVICE = 'removeAddTorService',
-  RESPONSE_GET_PUBLIC_CHANNELS = 'responseGetPublicChannels'
+  RESPONSE_GET_PUBLIC_CHANNELS = 'responseGetPublicChannels',
+  RESPONSE_GET_AVAILABLE_USERS = 'responseGetAvailableUsers',
+  RESPONSE_GET_PRIVATE_CONVERSATIONS = 'responseGetPrivateConversations',
+  RESPONSE_FETCH_ALL_DIRECT_MESSAGES = 'responseFetchAllDirectMessages'
 }