+++ packages/waggle/src/socket/events/channels.ts
@@ -1,7 +1,6 @@
-import { ChannelInfoResponse } from "../../storage/storage"
-import { EventTypesResponse } from "../constantsReponse"
+import { ChannelInfoResponse } from '../../storage/storage'
+import { EventTypesResponse } from '../constantsReponse'
 
 export const loadAllPublicChannels = (socket: any, channels: ChannelInfoResponse) => {
   socket.emit(EventTypesResponse.RESPONSE_GET_PUBLIC_CHANNELS, channels)
 }
-