+++ packages/waggle/src/torManager/torControl.ts
@@ -27,7 +27,6 @@ class TorControl {
 
     this.connect = function connectTorControl(params, cb) {
       params = params || opts
-
       if (this.connection) {
         if (cb) {
           return cb(null, this.connection)
@@ -45,9 +44,11 @@ class TorControl {
       }
 
       this.connection = net.connect(params)
+      console.log(this.connection)
 
-      //Handling connection errors
       this.connection.once('error', function (err: any) {
+        console.log('error connecting')
+        self.connection = null
         if (cb) {
           cb(new Error('Error connecting to control port: ' + err))
         }
@@ -55,9 +56,11 @@ class TorControl {
 
       // piping events
       this.connection.on('data', (data: any) => {
+        console.log(`data is ${data}`)
         self.eventEmitter.emit('data', data)
       })
       this.connection.on('end', () => {
+        console.log('connection ended')
         self.connection = null
         self.eventEmitter.emit('end')
       })
@@ -103,31 +106,35 @@ class TorControl {
       return this
     }
   }
-  private connect: (params: any, cb: any) => { any: any }
+
+  private readonly connect: (params: any, cb: any) => { any: any }
   private connection: any
-  private disconnect: any
-  private isPersistent: any
-  private setPersistent: any
-  private eventEmitter: any = new EventEmitter()
-  private sendCommand: any = (
+  private readonly disconnect: any
+  private readonly isPersistent: any
+  private readonly setPersistent: any
+  private readonly eventEmitter: any = new EventEmitter()
+  private readonly sendCommand: any = async (
     command: string,
     keepConnection: boolean
-  ): Promise<{ code: number; messages: string[] }> => {
-    return new Promise((resolve, reject) => {
-      var self = this,
-        tryDisconnect = function (callback: any) {
-          if (keepConnection || self.isPersistent() || !self.connection) {
-            return callback()
-          }
-          return self.disconnect(callback)
+  ): Promise<{ code: number, messages: string[] }> => {
+    return await new Promise((resolve, reject) => {
+      var self = this
+      var tryDisconnect = function (callback: any) {
+        if (keepConnection || self.isPersistent() || !self.connection) {
+          return callback()
         }
+        return self.disconnect(callback)
+      }
       return this.connect(null, function (err: any, connection: any) {
+        console.log('entered this.connect')
         if (err) {
+          console.log('reveived error inside this.connect')
+          console.log(err)
           return reject(err)
         }
         connection.once('data', function (data: any) {
           return tryDisconnect(function () {
-            let messages = []
+            const messages = []
             let arr = []
             data = data.toString()
             console.log('dataaa', data)
@@ -152,19 +159,24 @@ class TorControl {
     })
   }
 
-  public async addOnion(request: string): Promise<{ code: number; messages: string[] }> {
+  public async addOnion(request: string): Promise<{ code: number, messages: string[] }> {
     return this.sendCommand('ADD_ONION ' + request)
   }
-  public async delOnion(request: string): Promise<{ code: number; messages: string[] }> {
+
+  public async delOnion(request: string): Promise<{ code: number, messages: string[] }> {
     return this.sendCommand('DEL_ONION ' + request)
   }
-  public signal(signal: string): Promise<{ code: number; messages: string[] }> {
+
+  public async signal(signal: string): Promise<{ code: number, messages: string[] }> {
+    console.log('received signal')
     return this.sendCommand('SIGNAL ' + signal)
   }
-  public signalReload(): Promise<{ code: number; messages: string[] }> {
-    return this.signal('RELOAD')
+
+  public async signalReload(): Promise<{ code: number, messages: string[] }> {
+    return await this.signal('RELOAD')
   }
-  public async setConf(request: string): Promise<{ code: number; messages: string[] }> {
+
+  public async setConf(request: string): Promise<{ code: number, messages: string[] }> {
     return this.sendCommand('SETCONF ' + request)
   }
 }