+++ packages/waggle/src/utils.ts
@@ -1,13 +1,14 @@
 import fs from 'fs'
 
 export function createPaths(paths: string[]) {
-    for (const path of paths) {
-      if (!fs.existsSync(path)) {
-        fs.mkdirSync(path, { recursive: true })
-      }
+  for (const path of paths) {
+    if (!fs.existsSync(path)) {
+      fs.mkdirSync(path, { recursive: true })
     }
   }
+}
 
 export function fetchAbsolute(fetch: Function): Function {
-  return (baseUrl: string) => (url, ...otherParams) => url.startsWith('/') ? fetch(baseUrl + url, ...otherParams) : fetch(url, ...otherParams)
+  return (baseUrl: string) => (url, ...otherParams) =>
+    url.startsWith('/') ? fetch(baseUrl + url, ...otherParams) : fetch(url, ...otherParams)
 }